+++ svelte/src/components/Dialog.svelte
@@ -24,9 +24,15 @@ it under the terms of the GNU Affero General Public License.
     element.addEventListener('aui-close', handler)
     return () => element.removeEventListener('aui-close', handler)
   })
+  $: if (element) {
+    element.open = open
+    element.title = title
+    element.description = description
+    element.closeLabel = closeLabel
+  }
 </script>
 
-<aui-dialog bind:this={element} {open} {title} {description} {closeLabel}>
+<aui-dialog bind:this={element}>
   <slot />
   <span slot="footer" style="display:contents"><slot name="footer" /></span>
 </aui-dialog>